Adds psionics to character select part II
Fixes formatting issues with the new psionics menu section on the character setup screen.

Adds psionics whitelists, created in the same way as species whitelists with 'psionics' as the species field.

Psions created through whitelists are able to select either Latent, Operant or Master in any of the four available faculties, but are not able to be multi-faculty.